refactor(utils): split utilityObsidian god file into cohesive modules
The 1320-line src/utilityObsidian.ts mixed eight unrelated concerns in one
"junk drawer" module. Move each concern into its own focused src/utils/
module and keep utilityObsidian.ts as a thin (78-line) re-export barrel so
the established import surface keeps working unchanged.

New modules:
- utils/templaterIntegration.ts  Templater bridging + file-settle waits
- utils/editorInsertion.ts        cursor/link insertion in the active editor
- utils/templateFolderUtils.ts    template path/extension/folder resolution
- utils/fileOpening.ts            opening files + workspace-leaf selection
- utils/userScript.ts             loading user scripts (.js / ```js fences)
- utils/vaultQueries.ts           folder/file/tag lookups over the vault
- utils/obsidianCommands.ts       command-registry helpers
- utils/dates.ts                  date / natural-language-date helpers
- utils/excludeKeys.ts            generic object helper
- getChoiceType moved into utils/choiceUtils.ts

Behavior-preserving: public export surface is byte-identical (44 symbols,
verified by diff). tsc/lint/build clean, 2127 tests green, plugin loads with
no errors, and a template+date+value run verified end-to-end in the dev vault.
