+++ src/utils/vaultQueries.ts
@@ -1,0 +1,78 @@
+import type { App, CachedMetadata, TFile } from "obsidian";
+import { TFolder } from "obsidian";
+
+export function getAllFolderPathsInVault(app: App): string[] {
+	return app.vault
+		.getAllLoadedFiles()
+		.filter((f) => f instanceof TFolder)
+		.map((folder) => folder.path);
+}
+
+export function isFolder(app: App, path: string): boolean {
+	const abstractItem = app.vault.getAbstractFileByPath(path);
+
+	return !!abstractItem && abstractItem instanceof TFolder;
+}
+
+export function getMarkdownFilesInFolder(app: App, folderPath: string): TFile[] {
+	return app.vault
+		.getMarkdownFiles()
+		.filter((f) => f.path.startsWith(folderPath));
+}
+
+function getFrontmatterTags(fileCache: CachedMetadata): string[] {
+	const frontmatter = fileCache.frontmatter;
+	if (!frontmatter) return [];
+
+	// You can have both a 'tag' and 'tags' key in frontmatter.
+	const frontMatterValues = Object.entries(frontmatter);
+	if (!frontMatterValues.length) return [];
+
+	const tagPairs = frontMatterValues.filter(([key, value]) => {
+		const lowercaseKey = key.toLowerCase();
+
+		// In Obsidian, these are synonymous.
+		return lowercaseKey === "tags" || lowercaseKey === "tag";
+	});
+
+	if (!tagPairs) return [];
+
+	const tags = tagPairs
+		.flatMap(([key, value]) => {
+			if (typeof value === "string") {
+				// separator can either be comma or space separated
+				return value.split(/,|\s+/).map((v) => v.trim());
+			} else if (Array.isArray(value)) {
+				return value as string[];
+			}
+		})
+		.filter((v) => !!v) as string[]; // fair to cast after filtering out falsy values
+
+	return tags;
+}
+
+function getFileTags(app: App, file: TFile): string[] {
+	const fileCache = app.metadataCache.getFileCache(file);
+	if (!fileCache) return [];
+
+	const tagsInFile: string[] = [];
+	if (fileCache.frontmatter) {
+		tagsInFile.push(...getFrontmatterTags(fileCache));
+	}
+
+	if (fileCache.tags && Array.isArray(fileCache.tags)) {
+		tagsInFile.push(...fileCache.tags.map((v) => v.tag.replace(/^#/, "")));
+	}
+
+	return tagsInFile;
+}
+
+export function getMarkdownFilesWithTag(app: App, tag: string): TFile[] {
+	const targetTag = tag.replace(/^#/, "");
+
+	return app.vault.getMarkdownFiles().filter((f: TFile) => {
+		const fileTags = getFileTags(app, f);
+
+		return fileTags.includes(targetTag);
+	});
+}
